Data stored in txt files
Data regarding energy demand and generation throughout the year. Data obtained using PVGIS and Pear Tree stats, and manually copied into txt files to be used in Python. Different files are used in different iterations of the energy model. As the model was developed, fewer of these txt files were used as it is an inefficient method. They have only been uploaded so that earlier iterations of the code can be run.